Update script.js
Instead of treating each individual conversion factor as its own variable, I'm treating the amount the user has of each amount of HP currency as its own variable and modifying what they have by the constant conversion factor. If I have not broken anything in the process simple arithmetic would return the final conversion. Then it's a matter of writing it to the site.